+++ drivers/cdc/ad7746/iio_ad7746.c
@@ -168,7 +168,7 @@ static int32_t ad7746_select_channel(void *device,
 	return delay;
 }
 
-static int ad7746_iio_read_raw(void *device, char *buf, size_t len,
+static int ad7746_iio_read_raw(void *device, char *buf, uint32_t len,
 				   const struct iio_ch_info *channel, intptr_t priv)
 {
 	struct ad7746_iio_dev *iiodev = (struct ad7746_iio_dev *)device;
@@ -230,7 +230,7 @@ static int ad7746_iio_read_raw(void *device, char *buf, size_t len,
 	return iio_format_value(buf, len, IIO_VAL_INT, 1, &value);
 }
 
-static int ad7746_iio_read_scale(void *device, char *buf, size_t len,
+static int ad7746_iio_read_scale(void *device, char *buf, uint32_t len,
 				     const struct iio_ch_info *channel, intptr_t priv)
 {
 	int32_t valt;
@@ -257,7 +257,7 @@ static int ad7746_iio_read_scale(void *device, char *buf, size_t len,
 	return iio_format_value(buf, len, valt, 2, vals);
 }
 
-static int ad7746_iio_read_offset(void *device, char *buf, size_t len,
+static int ad7746_iio_read_offset(void *device, char *buf, uint32_t len,
 				      const struct iio_ch_info *channel, intptr_t priv)
 {
 	struct ad7746_iio_dev *iiodev = (struct ad7746_iio_dev *)device;
@@ -270,7 +270,7 @@ static int ad7746_iio_read_offset(void *device, char *buf, size_t len,
 }
 
 static int ad7746_iio_write_offset(void *device, char *buf,
-				       size_t len, const struct iio_ch_info *channel, intptr_t priv)
+				       uint32_t len, const struct iio_ch_info *channel, intptr_t priv)
 {
 	struct ad7746_iio_dev *iiodev = (struct ad7746_iio_dev *)device;
 	struct ad7746_dev *desc = (struct ad7746_dev *)iiodev->ad7746_dev;
@@ -318,7 +318,7 @@ static int ad7746_iio_write_offset(void *device, char *buf,
 	return len;
 }
 
-static int ad7746_iio_read_samp_freq(void *device, char *buf, size_t len,
+static int ad7746_iio_read_samp_freq(void *device, char *buf, uint32_t len,
 		const struct iio_ch_info *channel, intptr_t priv)
 {
 	struct ad7746_iio_dev *iiodev = (struct ad7746_iio_dev *)device;
@@ -341,7 +341,7 @@ static int ad7746_iio_read_samp_freq(void *device, char *buf, size_t len,
 }
 
 static int ad7746_iio_write_samp_freq(void *device, char *buf,
-		size_t len, const struct iio_ch_info *channel, intptr_t priv)
+		uint32_t len, const struct iio_ch_info *channel, intptr_t priv)
 {
 	struct ad7746_iio_dev *iiodev = (struct ad7746_iio_dev *)device;
 	struct ad7746_dev *desc = (struct ad7746_dev *)iiodev->ad7746_dev;
@@ -370,7 +370,7 @@ static int ad7746_iio_write_samp_freq(void *device, char *buf,
 }
 
 static int ad7746_iio_read_samp_freq_avail(void *device, char *buf,
-		size_t len,
+		uint32_t len,
 		const struct iio_ch_info *channel, intptr_t priv)
 {
 	int32_t ret = 0;
@@ -392,7 +392,7 @@ static int ad7746_iio_read_samp_freq_avail(void *device, char *buf,
 	return ret;
 }
 
-static int ad7746_iio_read_calibscale(void *device, char *buf, size_t len,
+static int ad7746_iio_read_calibscale(void *device, char *buf, uint32_t len,
 		const struct iio_ch_info *channel, intptr_t priv)
 {
 	struct ad7746_iio_dev *iiodev = (struct ad7746_iio_dev *)device;
@@ -423,7 +423,7 @@ static int ad7746_iio_read_calibscale(void *device, char *buf, size_t len,
 	return iio_format_value(buf, len, IIO_VAL_INT_PLUS_MICRO, 2, vals);
 }
 
-static int ad7746_iio_write_calibscale(void *device, char *buf, size_t len,
+static int ad7746_iio_write_calibscale(void *device, char *buf, uint32_t len,
 		const struct iio_ch_info *channel, intptr_t priv)
 {
 	struct ad7746_iio_dev *iiodev = (struct ad7746_iio_dev *)device;
@@ -452,7 +452,7 @@ static int ad7746_iio_write_calibscale(void *device, char *buf, size_t len,
 	}
 }
 
-static int ad7746_iio_read_calibbias(void *device, char *buf, size_t len,
+static int ad7746_iio_read_calibbias(void *device, char *buf, uint32_t len,
 		const struct iio_ch_info *channel, intptr_t priv)
 {
 	struct ad7746_iio_dev *iiodev = (struct ad7746_iio_dev *)device;
@@ -470,7 +470,7 @@ static int ad7746_iio_read_calibbias(void *device, char *buf, size_t len,
 	return iio_format_value(buf, len, IIO_VAL_INT, 1, &value);
 }
 
-static int ad7746_iio_write_calibbias(void *device, char *buf, size_t len,
+static int ad7746_iio_write_calibbias(void *device, char *buf, uint32_t len,
 		const struct iio_ch_info *channel, intptr_t priv)
 {
 	struct ad7746_iio_dev *iiodev = (struct ad7746_iio_dev *)device;
@@ -488,7 +488,7 @@ static int ad7746_iio_write_calibbias(void *device, char *buf, size_t len,
 	return ad7746_set_cap_offset(desc, (uint16_t)val);
 }
 
-static int ad7746_start_offset_calib(void *device, char *buf, size_t len,
+static int ad7746_start_offset_calib(void *device, char *buf, uint32_t len,
 		const struct iio_ch_info *channel, intptr_t priv)
 {
 	int32_t ret;
@@ -510,7 +510,7 @@ static int ad7746_start_offset_calib(void *device, char *buf, size_t len,
 	return len;
 }
 
-static int ad7746_start_gain_calib(void *device, char *buf, size_t len,
+static int ad7746_start_gain_calib(void *device, char *buf, uint32_t len,
 				       const struct iio_ch_info *channel, intptr_t priv)
 {
 	int32_t ret;
